+++ apps/react-chat/src/app/chat/core/providers/twitch/__test__/tmi-js-mock.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/ban-types
 let listeners: { [event: string]: Function } = {};
 
 export function triggerListener(event: string, ...args: unknown[]): void {
@@ -7,6 +8,7 @@ export function triggerListener(event: string, ...args: unknown[]): void {
 export const mockTmiJsConstructorCalls = jest.fn();
 export const mockTmiJsInstanceMock = {
   connect: jest.fn(),
+  // eslint-disable-next-line @typescript-eslint/ban-types
   on: jest.fn().mockImplementation((event: string, callback: Function) => {
     listeners = { ...listeners, [event]: callback };
   }),
