+++ apps/react-chat/src/app/chat/core/providers/twitch/blocks/smile/smile-formatter.spec.ts
@@ -1,5 +1,5 @@
-import { TwitchProviderMeta } from '../../providers/twitch/twitch.interfaces';
-import { ChatMessageBlock } from '../message-blocks.interfaces';
+import { ChatMessageBlock } from '../../../../message-blocks/message-blocks.interfaces';
+import { TwitchProviderMeta } from '../../twitch.interfaces';
 import { formatSmileBlocks } from './smile-formatter';
 
 describe('chat/core/message-blocks smile', () => {
