+++ apps/react-chat/src/app/chat/core/providers/twitch/blocks/twitch-blocks.consts.ts
@@ -1,4 +1,4 @@
 import { formatLinksBlocks } from './link/link-formatter';
 import { formatSmileBlocks } from './smile/smile-formatter';
 
-export const DEFAULT_FORMATTERS = [formatSmileBlocks, formatLinksBlocks];
+export const BLOCKS_FORMATTERS = [formatSmileBlocks, formatLinksBlocks];
