+++ apps/react-chat/src/app/chat/core/providers/twitch/use-twitch-chat.ts
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 import tmi from 'tmi.js';
 import { ChatMessage } from '../../../chat.interfaces';
 import { formatMessageTextToBlocks } from '../../message-blocks/message-blocks';
+import { BLOCKS_FORMATTERS } from './blocks/twitch-blocks.consts';
 import { TWITCH_PROVIDER } from './twitch.consts';
 
 export interface TwitchChatConfig {
@@ -30,6 +31,7 @@ export function useTwitchChat({ channel }: TwitchChatConfig) {
       const blocks = formatMessageTextToBlocks({
         text: message,
         meta: { tags },
+        formatters: BLOCKS_FORMATTERS,
       });
 
       const chatMessage: ChatMessage = {
