+++ apps/react-chat/src/app/chat/message/message-link-block/message-link-block.tsx
@@ -1,6 +1,6 @@
 import { styled } from '@mui/material';
 import { memo, useEffect, useState } from 'react';
-import { ChatMessageLinkBlock } from '../../core/message-blocks/link/link.interfaces';
+import { ChatMessageLinkBlock } from '../../core/providers/twitch/blocks/link/link.interfaces';
 
 export interface MessageProps {
   block: ChatMessageLinkBlock;
